requirements.txt
affine==2.3.0
albumentations==0.3.3
attrs==19.2.0
buzzard==0.6.4
certifi==2019.9.11
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
GDAL==2.4.2
gitdb2==2.0.6
GitPython==3.0.3
GPUtil==1.4.0
icecream==1.5.0
imageio==2.6.1
imgaug==0.2.6
joblib==0.14.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.2.0
Pint==0.9
protobuf==3.10.0
Pygments==2.4.2
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
rasterio==1.1.0
Rtree==0.8.3
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
Shapely==1.6.4.post2
six==1.12.0
smmap2==2.0.5
snuggs==1.4.7
sortedcontainers==2.1.0
tensorboardX==1.9
torch==1.2.0
torchvision==0.4.0
tqdm==4.36.1
untokenize==0.1.1
